etl/requirements.txt
psycopg2-binary
jsonschema
pytest
